apps/mcp_process_cccd/requirements.txt
mcp[cli]
playwright
google-generativeai
python-dotenv
